BUGFIX: Allow to delete "used" resources from a storage
This solves the following case…

Given these settings:

    resource:
      collections:
        readableFilenames:
          storage: 'readableFilenameResourcesStorage'
          target: 'readableFilenameResourcesTarget'
      storages:
        readableFilenameResourcesStorage:
          storage: 'TYPO3\Flow\ResourceManagement\Storage\WritableFileSystemStorage'
          storageOptions:
            path: '%FLOW_PATH_DATA%Persistent/ReadableResources/'
      targets:
        readableFilenameResourcesTarget:
          target: 'Acme\AcmeCom\FilenameFileSystemSymlinkTarget'
          targetOptions:
            path: '%FLOW_PATH_WEB%Files/'
            baseUri: 'Files/'

I want to "move" a resource from the `persistent` to the `readableFilenames`
collection. To do this, I get an asset, fetch the resource and import it
into the `readableFilenames` collection. After that the newly imported
resource is published, assigned to the asset and then the old resource is
deleted. Code would be something like this:

    $resource = $asset->getResource();

    $importedResource = $resourceCollection->importResource($resource->getStream());
    $importedResource->setFilename($resource->getFilename());
    $importedResource->setMediaType($resource->getMediaType());
    $resourceCollection->getTarget()->publishResource($resource, $resourceCollection);

    $asset->setResource($importedResource);
    $this->assetRepository->update($asset);

    $this->resourceManager->deleteResource($resource);

But this leads to log messages about the storage data not being deleted,
because the resource is still being used. Which is not true, or at least
not fully correct. The problem at this point: the same resource exists
in two collections, but the check only looks at the SHA1 (and filename,
partly).

So this change adjusts the checks involved to look at the collection a
resource is in, too.

+++ TYPO3.Flow/Tests/FunctionalTestCase.php
@@ -435,13 +435,22 @@ abstract class FunctionalTestCase extends \TYPO3\Flow\Tests\BaseTestCase
      */
     protected function cleanupPersistentResourcesDirectory()
     {
-        $settings = self::$bootstrap->getObjectManager()->get(\TYPO3\Flow\Configuration\ConfigurationManager::class)->getConfiguration(ConfigurationManager::CONFIGURATION_TYPE_SETTINGS);
-        $resourcesStoragePath = $settings['TYPO3']['Flow']['resource']['storages']['defaultPersistentResourcesStorage']['storageOptions']['path'];
-        if (strpos($resourcesStoragePath, FLOW_PATH_DATA) === false) {
-            throw new \Exception(sprintf('The storage path for persistent resources for the Testing context is "%s" but it must point to a directory below "%s". Please check the Flow settings for the Testing context.', $resourcesStoragePath, FLOW_PATH_DATA), 1382018388);
-        }
-        if (file_exists($resourcesStoragePath)) {
-            Files::removeDirectoryRecursively($resourcesStoragePath);
+        $settings = self::$bootstrap->getObjectManager()->get(ConfigurationManager::class)->getConfiguration(ConfigurationManager::CONFIGURATION_TYPE_SETTINGS);
+        foreach ($settings['TYPO3']['Flow']['resource']['storages'] as $storageName => $storageSettings) {
+            if (!isset($storageSettings['storageOptions']['path'])) {
+                continue;
+            }
+
+            $resourcesStoragePath = $storageSettings['storageOptions']['path'];
+            if (strpos($resourcesStoragePath, FLOW_PATH_DATA) === false) {
+                throw new \Exception(sprintf('The storage path for persistent resources for the Testing context is "%s" for the "%s" storage, but it must point to a directory below "%s". Please check the Flow settings for the Testing context.', $resourcesStoragePath, $storageName, FLOW_PATH_DATA), 1382018388);
+            }
+            if (strpos($resourcesStoragePath, '/Test/') === false) {
+                throw new \Exception(sprintf('The storage path for persistent resources for the Testing context is "%s" for the "%s" storage, but it must contain "/Test/". Please check the Flow settings for the Testing context.', $resourcesStoragePath, $storageName), 1382018388);
+            }
+            if (file_exists($resourcesStoragePath)) {
+                Files::removeDirectoryRecursively($resourcesStoragePath);
+            }
         }
     }
